Add files via upload
Energy Calculation:-Reducing energy consumption is essential for mitigating climate change, conserving natural resources, and reducing
energy costs. • Co-ordinates of vessel:-It is important to know the exact location
for optimizing the route. • Functionality of parts:-To get the predictive analysis of maintainence of
the vessel we need to know the functionality of the parts. • EAT(Estimated Arrival Time):-It is used for identifying the best route by
taking origin, destination and giving EAT.